pyproject.toml
[tool.poetry]
name = "hypoevolve"
version = "0.5.0"
description = "A simplified and efficient implementation of the AlphaEvolve mechanism"
authors = ["Assistant <assistant@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.90.0"
pyyaml = "^6.0.2"
numpy = "<2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
